third_party/blink/web_tests/wpt_internal/css/css-flexbox/balance/flex-line-count-invalid.html

<!DOCTYPE html>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/3070">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<body>
<script>
test_invalid_value("flex-line-count", "1.5");
test_invalid_value("flex-line-count", "0");
test_invalid_value("flex-line-count", "-1");
test_invalid_value("flex-line-count", "auto");
test_invalid_value("flex-line-count", "wrap");
</script>
</body>
